# Inline System.Boolean Parse helper methods