- Profiling the boolean parse method showed that its
  helper methods weren't being inlined
- Increases performance on the System.Tests.Perf_Boolean.Parse*
  microbenchmarks