@L2 2022-02-03
summary:
better: 2, geomean: 7.583
total diff: 2

No Slower results for the provided threshold = 5% and noise filter = 0.3ns.

| Faster                                          | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Tests.Perf_Boolean.Parse(value: "True")  |      7.97 |            21.99 |             2.76 |         |
| System.Tests.Perf_Boolean.Parse(value: "False") |      7.22 |            21.96 |             3.04 |         |