@danmoseley 2022-02-03
Thanks. It's unfortunately hard to make inlining decisions by running microbenchmarks - the microbenchmarks almost invariably get faster, but in a real app, the code can get significantly bigger.

If it's helpful to inline these, perhaps we should consider why they're not getting inlined.