@stephentoub 2022-02-04
The implementations can be streamlined with fewer reads:
https://github.com/dotnet/runtime/pull/64782