@L2 2022-02-04
> Thanks. It's unfortunately hard to make inlining decisions by running microbenchmarks - the microbenchmarks almost invariably get faster, but in a real app, the code can get significantly bigger.
> 
> If it's helpful to inline these, perhaps we should consider why they're not getting inlined.

Sounds good, I'll close this PR for now as @stephentoub 's implementation is the better way to go to extract more performance. Thanks again @danmoseley 

> The implementations can be streamlined with fewer reads: #64782

@stephentoub , very nice!