@stephentoub 2022-02-04
> And we appreciate the inspiration here

Yeah I wouldn't have noticed if this PR didn't draw my eyes to it :-)